backend/test_service/requirements.txt
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
psycopg2==2.9.9
python-dotenv==1.0.1
gunicorn==23.0.0